All window views are now created so they are as large as they need to be to fully display their complete contents.
On layout all views within a workspace have their borders limited so they dispay within that workspace.

In addition to the above limits being imposed the browser frame will reduce the size of the right hand pane so it is accommodated by the reduced frame.

Laying out the browser frame so that the left and right pane are the same height, accommodating all the content in each pane.

Forced the spy window to be hidden.  It is still updating the detials, just not showing it.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006477 13f79535-47bb-0310-9956-ffa450edef68