Limited the bounds of the overlay views so they can't be dragged out of the frame.
Limited the bounds of the menu so it alway appears with all options completely shown.

Added scroll bars to the error view so all the error text can be viewed.

The spy window can now be closed and reopened from the debug menu.  When closed all call to it are ignored.

Started tidying up table browser.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006484 13f79535-47bb-0310-9956-ffa450edef68