Added methods to ActionAbout interface to allow parameter labels and default values to be set for parametered action methods.
Updated the reflector mechanism to determine what interfaces are used by a naked object.

Modified the isOfType method in NakedObjectSpecificationImpl to check to see if the specified type is one of the interfaces in the declaration.  Prior to this only the classes where being checked.

All vector collection now throw IllegalArgumentException when the contains method is called with null, as it is not object that the collection can ever contain.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006485 13f79535-47bb-0310-9956-ffa450edef68