Fixed bug where dates where not comparing correctly, after being cleared and edited.
Fix bug in TestObjectImpl where the associate methods where not first checking the abouts.

Added build number to the .NET build when building the release.

Double-clicking an icon now opens up a new top-level view of that object.

Created ValueMemento to allow value objects to be converted to string data and recreated at some point in the future.

Made all the restore and save methods in the NakedValue classes deal with empty values consistently, clearing the restored values if they data is null.

Modified all the invoke action methods in TestObject to take a TestNaked rather than a TestObject in the single parameter form.

Made all objects used to transport information during distribution implement the Serializable interface so they can be serialized.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006504 13f79535-47bb-0310-9956-ffa450edef68