Devised a mechanism for server-collabrative instance searches.
Prevented non-persistent objects from being used to set up fields in persistent objects.

Fixed bug where objects dropped onto lists in trees and forms where not being added to the collection.

Fixed bug where the logging of update messages caused an error and prevented actions being completed.

Fixed bug in invokeAction in TestObjectImpl where the about was not being called first, and hence was being allowed to be invoked when it shouldn't be.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006515 13f79535-47bb-0310-9956-ffa450edef68