Disallowing aggregated objects (as marked by the Aggregated interface) from being dropped into fields when the source object already has an owner.
Fixed baseline on empty field to be constant based on the size of the font used for the label.  This was making the rows in table misalign.

Table column name are now being truncated so they do not run into the next column.

Column headers are now looking for the name specified by the about, if there is one.

Stopped the layout mechanism from repeatedly running when the available size was less than the required size.

Created column width strategy interface for determining the width of columns.  Implemented a default version that sizes each columns according to its type.

Fixed dialog boxes so that the view status, specifically the red/green flashing, shows.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006519 13f79535-47bb-0310-9956-ffa450edef68