Added the invoke methods offered by TestObject to TestClass, allowing class methods to be tested and run.
Refactor TestObject and TestClass so that all the invoke method are provided by a superclass.

Replaced the action methods that took a single TestObject parameter with methods that take a TestNaked; now there are zero, one, and many parameter versions of each method, and each one can take a test value or object.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006521 13f79535-47bb-0310-9956-ffa450edef68