Fixed bug with dialog lookups where the field was no longer allowing the instances to be shown.
Changed the default invalid colour to a more suitable one for use as backgrounds behind text fields (when invalid entry is made).

Modified the the selection overlay view (as used for lookups) so that it has it bounds limited by the application window.  Also added a scroll border to it so that if the list is too big to fit on the screen then the list will be scrollable.

Fixed the repeated laying out of a view when it is too big to fit on the application window, which caused a performance problem.  Now the view is laid out once.

Removed the 'clear collection' option from list/tables.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006549 13f79535-47bb-0310-9956-ffa450edef68