Made the actions "instances" and "new instance" exploration actions.
Now running actions as background tasks.  To indicate this to user the border changes color (to red) while the action is executing.

Rearranged keyboard modifiers so that the debug and exploration menus are showing.  Alt is now used to access the alternative menu (the view menu when you would normally get the object menu, and the object menu when you would normally get the view menu).  Ctrl now accesses the exploration menu, and Ctrl+Shift accesses the debug menu.

Transient object now push for persistence by offering a "save", a "save and close" and a "discard" button on their views. Additionally the close button on the title bar is not shown.  When the either of the save buttons are pressed the viewer will invoke a method on the object called "save" or "persist", or if neither are available, will make it persistent via the object manager.

The classAbout mechanism now works up the class hierachy.

The viewer's quit mechanism no longer shuts down the system.  The client program is now expected to register a ObjectViewingMechanismListener and respond to it by shutting the system down.

Fixed a problem where class methods where not be recognised, by name, in XATs for invoke and assert requests.

View suitable is now being done by passing the Content object to the view specifications, and not just the object.  This allows for greater control on what views are available to show specific objects in various contexts.  Specifically, tables are now been made available where are suitable.

The window title border will now respond to double-click to minimize them to an icon.  Also, an option has be been added to the menu to give the same result.

In AcceptanceTestCase the getForObject now returns the business object and the not wrapping object, ie the NakedObject.  To get the NakedObject use getForNakedObject().

The buttons "Execute" and "Execute and Close" have been renamed to "Apply" and "OK".

The option "Make persitent" has been removed from object views.  Any need to make a transient object persistent will now need to have a explicit method within the class.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006691 13f79535-47bb-0310-9956-ffa450edef68