Tidied up Naked, NakedObject and NakedObjectSpecification.  Removed AbstractNakedObject, moving all necessary code into PojoAdapter.
Refined the possible states of an object - to cover: transient, unresolved, partly resolved, resolving and resolved.

Added a notification method to Reflector to indicate the destroying of an object.  This should be passed on to the reflector where the POJO should be notified.

Added asserts that can take a message and a target object.  This allows details about the target to be output when an assertion fails, without having to concatenate the object with the message.

Consolidated the Dummys and Mocks for NakedObject and NakedObjectSpecification I/Fs.

Fixed the remaining broken tests.




git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1007201 13f79535-47bb-0310-9956-ffa450edef68