ISIS-14: fixing support for lazily loading.
Collections that are lazily loaded were not being synchronized with the framework
(ie it thought that the pojos were transient).

This fix extends the PojoRecreator interface to also be able to detect that
a pojo has actually been lazily loaded previously.

At the same time, have also improved the logging of the IsisLifecycleListener, and
have consolidated all the logic that pertains to keeping JDO and Isis in
sync into a new FrameworkSynchronizer component.  This is delegated to by
the IsisLifecycleListener, the ObjectStore and the PojoRecreator.

git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1378595 13f79535-47bb-0310-9956-ffa450edef68