ISIS-351: improving exception handling...
... looking into the case when hit an DB exception (unique constraint violation) on an edit.
Noticed that the version was being bumped anyway even though change not actually committed to DB.
After some experimentation, determined that if the transaction is aborted during the first phase,
then the version is not bumped up, but if the transaction is aborted during the second phase (by inspecting
for any "abort causes" on the IsisTransaction, then the version got bumped up.

Moved WebRequestCycleForIsis#onEndRequest logic instead into the #onRequestHandlerExecuted callback
which is executed before

Also reworked and simplified a little of the IsisTransaction and IsisTransactionManager code (in particular
the IsisTransactionManager#endTransaction method).