ISIS-599: config property to enable logging in exception recognizers
Thus, if "isis.services.exceprecog.logRecognizedExceptions" set to true,
then a stack trace of any recognized exceptions will be logged.

If the exception is NOT recognized, then no stack trace is logged; the
assumption is that if none of the recognizers recognize the exception,
then it will bubble up and its stack trace be logged anyway as a
unexpected exception.

In addition:
- simplified the ExceptionRecognizer hierarchy
- made ExceptionRecognizers be init/shutdown via @PostConstruct methods etc
  (to read the configuration)
- updated the example isis.properties files for the archetypes.