ISIS-651: fixing NPE in ActionMemento and similar
Also: forcing invalidation of Isis ObjectSpec in EntityPage (else the UI is rendered w.r.t old Isis metadata)

Not exactly sure why either of these are problems.  For the latter, certainly, it seems that the IsisJRebelPlugin invalidates the
cache correctly, but it isn't picked up.  It would seem that the Wicket viewer is holding the ActionMemento, PropertyMemento and CollecitonMemento in-memory?

also:
- adding maxPermSize to launch config for simple app.