ISIS-760: don't create an xactn for service adapters...
... this is a left-over of an earlier design.

The refactoring done in earlier commits for this issue causes the endTransaction to do a close() on all services, and in the case of any services that register with EventBusService (eg ToDoItemSubscriptions), this call to close() happens before an open().  We end up with a stack-trace and non-bootable system.

This smacks of an assymetry here that probably should be investigated further (should beginTransaction do an open on all services if endTransaction is doing a close?) but in any case the quick fix is simply avoid creating the xactn anyway for the service adapters, as there's no need.