ISIS-537: pattern matching of CSS classes based on action name (similar to FA icons)
Also: required to inject ServicesInjector into facet factory to query whether a method was from a service or not.  To support this, refactored so that ServicesInjectorDefault is basically has global scope as opposed to session/request scope.  Implementation adapted from that already used for SpecificationLoader (aka the ObjectReflector).  A bit hacky, but a move in the right general direction.