ISIS-1044: when filtering collections for visibility, ensure resultant object is of the same type as the original.
as per mailing list (http://isis.markmail.org/thread/6pihpktvp7zly5vn) ... previously was always returning an ArrayList, which would result in a ClassCastException if the actual type was something else, eg a SortedSet (and called via wrapper factory... doesn't show up as an issue if just using generic UI).

The implementation suggested by Nacho in the ML would be ok for action invocations (a standalone list) but might result in issues if manipulating the collection returned by DataNucleus for a "managed" internal collection.

So have preserved the intent of the original design, but have introduced CollectionUtils#copyOf(...), that copies the collection but also copies to the expected type.   If cannot copy, then returns null (and original unfiltered collection is returned instead).