ISIS-1044: filtering changed so that it is "opt-in", only enabled from selected callers.
It turns out that it's wrong to perform filtering all the time at the facet (metamodel) level... there are other internal "trusted" callers of the metamodel for which filtering should not be performed.  Examples are the auditing support, and creating XmlSnapshots.

As a first step to refactoring,  have put a guard around the code via a thread-local, and provided an API - InteractionUtils.withFiltering(Callable) -  which callers can use when they do want filtering.

Have then gone through the Wicket viewer (EntityModel, EntityCollectionModel, ActionModel) and also the WrapperFactory (DomainObjectInvocationHandler) to enable:
- In the case of the wicket viewer, it is enabled for property references, collections, action results and also  autocomplete, choices.
- For the wrapper factory, it is enabled provided that the execution mode is NOT skipRules.

Note that the Restful Objects viewer currently has NOT been updated.

In addition, have simplified the configuration properties to a single property, "isis.reflector.facet.filterVisibility".

~~~
All that said, I'm now realizing that even this isn't quite safe enough... the thread-local stays in effect for the duration of the call to the getter or action or autoComplete/choices method etc, and that in turn could call domain services or flush the transaction which results in framework-level code (eg auditing) then running while the thread-local is in effect.  So next step is to get rid of the thread-local just introduced, and simply filter in the caller afterwards.  I'll get there on this one, eventually...