ISIS-1239, ISIS-1324, ISIS-1328: multi-select for standalone collections was broken, now fixed.
Issue was that earlier work moved entire responsibility for being the "UiHintContainer" onto the EntityModel, on the basis that only entities can store hints over time.  However, the selector drop-down for standalone collections also was storying the hint on the UiHintContainer, which previously was also supported by (standalone) EntityCollectionModel.  This meant that the hint got lost.  Fix for that is to allow the IsisUiHintEvent to also directly hold the hint info.

In addition, the algorithm for selecting the default view was broken for standalone collections; should default to "table" if a standalone and there is no other metadata available.