ISIS-1441: removes AuthenticationManagerStandardForDfltRuntime, instead use AuthenticationManagerStandard.
The supporting classes move from isis-core-runtime to isis-core-metamodel.  Because the Authenticator subclasses relied on DeploymentType (only available in the isis-core-runtime), these instead depend on DeploymentCategory (in isis-core-metamodel).

This also includes a minor refactoring in the closeSession stuff... previously we called closeSession on AuthenticationManagerStandardForDfltRuntime, that in turn called IsisSessionFactory#closeSession() in order to clear the threadlocal holding the current session.  However, IsisSessionFactory is part of isis-core-runtime so this design doesn't work.  However, there is only one caller of closeSession, namely UserResourceServerside (the RO viewer).  The refactoring is therefore easy enough... just have UserResourceServerside make the call to IsisSessionFactory.