ISIS-1447: disables behaviour that (we believe) may cause 500 errors for Restful Objects server (Shiro authentication).
As per http://markmail.org/message/hsjljwgkhhrzxbrm, we believe that the auto-logout behaviour within the IsisSessionFilter for Restful Objects, if configured to use Shiro, may result in a 500 errors in certain race conditions.  This commit removes this auto-logout behaviour (simply continue to use the existing Shiro credentials).  The original behaviour can be reinstated using a new configuration parameter.