ISIS-1453: fixes SessionManagementService to call the correct API to set up the next xactn.
Use of the wrong API was resulting in the Isis framework's notion of the current transaction being out-of-whack with JDO's, resulting in the framework attempting to start the same JDO xactn twice.

In addition, fixed handling of closing the command when  first session is closed; was resulting in NPE because the interaction.Execution is still current and thus integration.getPriorExecution() returns null (resulting in an NPE).